+++ src/utils/helper/index.ts
@@ -2,6 +2,8 @@ export * from './address';
 export * from './balance';
 export * from './block';
 export * from './common';
+export * from './http';
+export * from './operator';
 export * from './storage';
 export * from './url';
-export * from './validate';
+export * from './validator';
