+++ src/utils/record/index.ts
@@ -1,3 +1,2 @@
-export * from './api';
 export * from './issuing';
 export * from './redeem';
